refactor(chatCore): extrai extractSystemRoleMessages para leaf puro (#3501)
Move a função top-level pura extractSystemRoleMessages (lift de mensagens
role=system/developer para o campo top-level `system`, exigido pela Messages API
da Anthropic) para o novo leaf open-sse/handlers/chatCore/claudeSystemRole.ts,
byte-idêntico. O chatCore.ts importa e **re-exporta** o símbolo, então os
importadores existentes (tests/unit/system-role-extraction.test.ts) seguem
resolvendo de chatCore.ts; os 3 call sites internos (incl. dentro de
normalizeClaudeUpstreamMessages) usam a versão importada — leaf autocontido, sem
import circular.

chatCore.ts encolhe ~33 linhas; typecheck 0; file-size OK. Coberto pelos 10
casos de tests/unit/system-role-extraction.test.ts (via re-export). complexity
1908=base v3.8.33 (drift pré-existente, mover função top-level é neutro).
